feat(ignore-file): Add support for preserving comments/blank lines (#2501)
Several project types and components in the codebase try to include comments in their ignore file entries, but the component was excluding them from the file. 

Comments are useful for users to understand the entries in their ignore files, especially entries that they did not add directly/first-hand.

End users can now customize what's included in the gitignore as part of the base project options:
```js
const p = new Project({
   name: 'my-project',
   gitIgnoreOptions: {
      filterCommentLines: false,
      filterEmptyLines: false,
   }
})
```

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.