fix(eslint): deprecate `lintProjenRcFile` in favor of explicitly defining rules via `Projenrc` (#2507)
Currently rules targeting the `Projenrc` file are spread between the `Eslint` and `Projenrc` components.
This is quite awkward to manage and also causes unncessary rules to be added for TypeScript Projenrc files and subprojects.

If you are currently using the `Eslint` component explicitly, we recommend to set `lintProjenRc: false` and add all required rules manually.

Also adds a feature `addLintPattern` to add lint patterns after the component has been created.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.