fix(github): extend PullRequestOptions to include PushOptions (#2511)
Hello 👋 firstly, thank you for such an intuitive and easy way to manage projects!

I found this slight issue whilst trying to create a workflow trigger to only occur when a PR is opened and changes occurred in a specific folder, but noticed this wasn't currently possible due to the API available.

The docs for `PushOptions` stated that the `pull_request` event can use these properties (and they are available to use as per the GitHub docs), but `PullRequestOptions` didn't extend `PushOptions`, instead only `PullRequestTargetOptions` did. This PR updates the doc string to include `pull_request_target` and makes `PullRequestOptions` extend `PushOptions`.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.