chore: fix test that updates snapshot locally
Replacing the external module with a different one.
The previous external module depends on an older version of projen that does not have a required bug fix.
Depending on system configuration, the test would either use the latest projen from the repo (this happens in CI/CD) or the outdated version (this happens locally).
Thus creating different snapshots.