chore: fix test that updates snapshot locally (#2518)
Replacing the external module with a different one. The previous external module depends on an older version of projen that does not have a required bug fix. Depending on system configuration, the test would either use the latest projen from the repo (this happens in CI/CD) or the outdated version (this happens locally). Thus creating different snapshots.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.