chore(javascript): fix invalid arg name in JSDocs (#2522)
Looks like I used an argument type instead of argument name in one spot of my recent PR #2515 in the JSDocs. I may look at building a test to catch this in the future but since it's Friday, here's a quick fix for the issue I created instead.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.