fix(subprojects): unnecessary use of `cd` when spawning default task of root project (#2542)
Instead of doing `cd` which is not as portable, we should use the existing feature to run commands from any directory.
This was an oversight when the feature was originally added.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.