feat(jsii): support for setting new `jsiiVersion` (#2539)
Now that jsii@5.0.0 is releases, we can allow users to explicitly set their jsii version.
For newer versions, we also no longer need to pin dependencies to versions compatible with TypeScript 3.9.

Closes #2264
Closes #2165

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.