fix: projen cannot be used with node "< 16.0.0" (#2547)
Reverts the accidental change of the minimum node version for projen itself. Node 14 is not EOL yet and we will only change the minimum node version a fair time after EOL.

Using projen with a version before Node 16 currently fails with an error like this:
```
error projen@0.69.1: The engine "node" is incompatible with this module. Expected version ">= 16.0.0". Got "14.21.3"
```

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.