fix(cli): command suggestion prevents use of external projects (#2557)
This changes keeps command suggestion, but adds a new "catch-all" subcommand to "projen new".
When we don't match a built-in project type, the "Invalid project type" error message is shown. That error could further be improved by adding some fuzzy string matching like yargs does, but for now this is good enough.

Fixes #2093

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.