feat: add `workflowContainerImage` to `Publisher` (#2572)
Currently the `workflowContainerImage` doesn't get propagated to the `package` or `release` jobs

This PR makes sure that in `jsii-project` when creating the `package-[language]` jobs we set `container` in `extraJobOptions` if present.

It also makes sure when instantiating a `Release` object we can pass it to `release_[language]` jobs.

Both of these jobs currently propagate `workflowRunsOn` to them so the precedent is there.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.