fix(new): cli fails on windows with RangeError (#2575)
The filesystem root check in the custom findUp implementation did not account for windows paths.
This was introduced in #2418 

Manually verified the fix on my windows machine.

This is blocking when a new project is created.
Existing projects were only partially affected, as the cli would have successfully found a projen directory unless run outside a projen dir. But there it would have errored anyway.

Fixes #2574


---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.