fix(file): marker contains wrong rcfile name (#2585)
Previously the rcfile path was detected at the point the target file component was added.
This resulted in an incorrect path if the Projenrc component was not yet added at this point.

The change moves the marker generation to synth time when all components are available.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.