feat(cli): validate options when creating projects from external packages (#2563)
Options validation for built-in project types is handled by yargs.
When using a project type from an external package via `--from`,  yargs is bypassed for options.
This meant that previously the validation and parsing behavior was incomplete.
Refactored code so that built-in and external types share much more codepaths.

Additionally align default value type checks in the inventory with options checks, fixes #2487

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.