fix(jest): cannot set `preset` and `globals` config when using in a typescript project (#2567)
Some of `jestConfig` properties are not able to update through constructor
because of a bug by assigning hard coded default values.

This commit fixes the bug to assign hard coded values only if the same
properties are not set.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.