feat(jsii): new `JsiiProject`s default to version `~5.0.0` (#2598)
Introduces a new custom doctag `@pjnew` to force an initial value at project creating time.
This will allow us to change defaults we previously couldn't because the impact would be to complicated.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.