fix(node): typesVersions is added by jsii  and removed by projen (#2603)
`jsii@5.x` introduced a compatibility feature that updates the `typesVersions` field in `package.json`.
If the field is added is dynamic and depends on the package. A future version of jsii might also not require compat and remove the field again.
Previously projen would remove the contents of this field at synth time.
With this change existing content is kept as is and overrides can be added if required.

Expose `file` property on `NodePackage` as it is practice with other Components that are wrapping a File component.

Additionally this change does some cleanup on how other fields are managed.
Particularly it removes the usage of `addOverride` from the Component as this might interfere with user overrides.

Finally the use of `NodeProject.hasScript()` has been deprecated. This was an oversight as it is simply a proxy to an already deprecated method.
Also deprecates `NodePackageOptions.scripts` which has an odd side effect of adding tasks to the project. This is also a left-over from when npm scripts and tasks were strongly coupled.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.