fix(release): incorrect prerelease version when there is a regular version (#2609)
As-is situation is when the following tags exist and when a new prerelease published, the nextVersion is selected as v1.0.1-pre.2 instead of v1.0.2-pre.0. This PR is to fix this behaviour.

- release v1.0.0
- release v1.0.1-pre.0
- release v1.0.1-pre.1
- release v1.0.1

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.