feat(java): expose readonly maven packaging dist dir (#2624)
It can be useful to know the location of the dist dir, eg. for setting up local dependencies in a monorepo (ref: https://github.com/aws/aws-prototyping-sdk/pull/358 )

This change exposes it as a readonly property.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.