feat(node-project): use resolution-mode=highest for pnpm by default
PNPM resolves dependencies to the lowest matching version by default, which diverges
from the more expected behaviour of NPM and Yarn which resolve to the latest matching
version.

Since projen uses '*' for dependencies where a version is not specified by the user,
PNPM resolves to the lowest published version for these which can cause problems,
for example ts-node resolves to 0.2.0 on initial project synthesis which causes
further calls to projen to fail.

Please note: `.npmrc` file is always created for node projects. Reference the
project's `npmrc` property to edit this instead of creating your own, or use
`project.tryRemoveFile('.npmrc')` first.

Fixes #2628