fix(javascript): do not rely on `package.json` being a valid entrypoint for dependency resolution. (#2645)
Fixes #2644

Instead of expecting `{package}/package.json` to always be a valid endpoint, this PR adds another strategy for resolving the manifest path.

If a lookup of `{package}/package.json` fails, this will now resolve the default entrypoint path of given module and search up the directory tree for the nearest `package.json` manifest file. If found and the manifest name matches the given dependency name, it will be used.

---

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.