fix(node-package): allow to remove resolutions (#2685)
This PR switches the source of truth for package resolutions from the internal structure `packageResolutions` to project dependencies (as in the case of all other dependencies).

This allows particularly to remove resolutions via simple `project.deps.removeDependency(name, DependencyType.OVERRIDE)` which is currently impossible (as resolution renders out of private `packageResolutions` and never syncs with the project dependencies after initial addition).

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.