feat(task): support adding condition to task (#2694)
Use Case:
- As a developer, I want to be able to add a condition to an existing task, so I can control how the task is executed during CI build

For instance, I want to add a condition to the "test" task, which is nested in the "build" task. During a CI release workflow, I want to skip tests during build, since they have already run in previous workflow job.

Currently to do this, I must reset the test task and recreate it.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.