feat(tasks): getter to return all env variables (#2765)
Currently environment variables can be defined in the task level but those variables cannot be retrieved, which is not the behaviour in the TaskStep level.

A new getter is added called `envVars` to retrieve the Task level environment variables. It is not possible to use `env` since there is already a functionall method `env(name:string, value:string)`.

Reroll of #2704 

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.