chore: use Node16 module & moduleResolution when appropriate
Instead of using `CommonJS` and `Node`, apply the `Node16` module and
module resolution settings, which are preferred when targeting node >=16
as it allows dynamically choosing the correct algorithms based on file
extension as well as package.json configuration.

Additionally, automatically configure `target` and `lib` based on the
configured minimum node version (if present).