feat(upgrade-dependencies)!: switch to single commands that only use `--filter` (#2845)
Currently, we leverage 3 separate features in order to support all various configurations that an upgrade workflow can accept:

- `--reject` is used to facilitate the [`exclude`](https://github.com/projen/projen/blob/main/src/javascript/upgrade-dependencies.ts#L36) property.
- `--filter` is used to facilitate the [`include`](https://github.com/projen/projen/blob/main/src/javascript/upgrade-dependencies.ts#L43) property.
- `--dep` is used to facilitate the [`types`](https://github.com/projen/projen/blob/main/src/javascript/upgrade-dependencies.ts#L97) property.

This, in combination of the fact that `peerDependencies` are recorded as `devDependencies` as well, and `bundledDependencies` are recored as regular `dependencies` as well, created to much snowflaking in the code.

Since projen has knowledge of all project dependencies, we can actually make it a lot simpler and drop the use of `--dep` and `--reject`. Instead, we only build `--filter` list that is relevant to the specific dependency types that are requested.
That way, we only have positive commands that are also much easier to reason through. 

BREAKING CHANGE: the `renderUpgradePackagesCommand` has been removed from `NodePackage`. It should never have been used. Instead, configure the appropriate `depsUpgradeOptions` to control the upgrade task. If you render the upgrade command for a different purpose, you'll have to construct the command yourself according to the appropriate package manager. 

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.