feat(github): require contributor statement on PRs (#2892)
Use`pull-request-lint` GitHub Actions Workflow to check if a required contributor statement has been included in the PR body. A list of users and labels can be provided to exempt certain PRs from this requirement.

[Tested on my fork.](https://github.com/mrgrain/projen/pull/2)

<img width="828" alt="image" src="https://github.com/projen/projen/assets/379814/bdb6f56f-b971-4367-aa2d-2fd1aeb5c17c">
<img width="874" alt="image" src="https://github.com/projen/projen/assets/379814/4b13bcf6-50d4-4d35-8f29-ae88da7b19fe">
<img width="664" alt="image" src="https://github.com/projen/projen/assets/379814/6f64f6ba-a098-4018-9683-dce8bab7b5b5">

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.