feat(docker-compose): Expose YAML file as property (#2866)
The docker compose components don't fully implement the compose spec, so there's often a need to directly extend/mutate the YAML file. Making the file accessible in the same way that it's done in a lot of other components.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.