feat(vscode): expose configuration files as `.file` property (#2998)
This PR exposes VSCode related JSON files as `.file` properties on appropriate components.

This will allow to use `.patch` and `.addToArray` method of ObjectFile for advanced customisations (as all those files are created lazily using `project.tryFindObjectFile` may not be a viable alternative currently).

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.