feat(gitlab)!: updates to configuration model (#2992)
Added the missing keywords for Services
* pull_policy: https://docs.gitlab.com/ee/ci/yaml/#servicepull_policy
* variables: Added in Gitlab 14.5 https://docs.gitlab.com/ee/ci/services/index.html#available-settings-for-services

BREAKING CHANGE: All `variables` are now Records of strings, as per documentation. If you are currently passing in a number, please convert it to a string.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.