feat(json-patch)!: each `ObjectFile.patch()` is an atomic operation and new failure behaviors for `JsonPatch.test()` (#2997)
`test` operation in JsonPatch are affecting ["patch as whole"](https://jsonpatch.com/#test) (`apply` operation of `fast-json-patch` actually throws when a test condition fails). However, `ObjectFile.patch` method are putting all patches into a single flatten array, discarding the concept of "operations" and actually disallowing use of `JsonPatch.test` in components (as it will broke whole file synthesis on a test failure).

This PR fixes the issue by considering all patches passed to a single call of `objectFile.patch` to be an atomic operation. That means either patches are applied, or none. When a `test` operation fails, only the current atomic operation fails. Any other will still be executed. This also changes the failure behavior of `JsonPatch.test` to fail silently, as opposed to the current exception. For this, we are also adding an optional parameter to `JsonPatch.test` to define the failure behaviour as either `SKIP` (new default behavior) or `FAIL_SYNTHESIS` (previous behavior). This allows to use JsonPatch for upsetting or conditionally changing some values based on current values (see added test):

Example:

```ts
 // adding object(s) if not exists - it's not possible to do something like this currently
objectFile.patch(
        JsonPatch.test("/first", undefined),
        JsonPatch.add("/first", {})
)
objectFile.patch(
        JsonPatch.test("/first/second", undefined),
        JsonPatch.add("/first/second", {})
)
objectFile.patch(
      JsonPatch.add("/first/second/field1", "value1"),
      JsonPatch.add("/first/second/field2", "value2"),
)
```

In additional to above changes this PR includes following complementary changes:
* updated version of `JsonPatch` that uses `deepClone` on patching operations
* exposing very useful `escapePathComponent` static method (imported from the same `fast-json-patch` dependency) to simplify operation on fields containing `/` and `~` (i.e. scoped dependencies overrides in `package.json`)

BREAKING CHANGE: A failure of `JsonPatch.test` does not fail synthesis anymore. To preserve previous behavior, pass `TestFailureBehavior.FAIL_SYNTHESIS` as third parameter to the `JsonPatch.test` call.

BREAKING CHANGE: Calls of `JsonPatch.test` are now only relevant to the scope of the list of patches passed to `ObjectFile.patch`. This should not break anything for truthy tests, and should not be relevant for failing tests. However, in case any issues are encountered, the previous behavior can be restored by placing all `JsonPatch`es in a single call of `ObjectFile.patch`.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.