chore: test "--git (default) respects init.defaultBranch setting" should not change user gitconfig
The test was inadvertently changing a user's local git config and was also not compatible with complex local git configs.

We can use env variables to simulate a specific git config without changing the actual configuration. This is okay because it's supported by git and has the same effect.

(Pulling out smaller changes from #2974 to make it easier to review)