chore: test "--git (default) respects init.defaultBranch setting" should not change user gitconfig (#3027)
The test was inadvertently changing a user's local git config and was also not compatible with complex local git configs.

We can use env variables to simulate a specific git config without changing the actual configuration. This is okay because it's supported by git and has the same effect.

(Pulling out smaller changes from #2974 to make it easier to review)

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.