fix(node): NpmConfig component is created even when not used (#3028)
By default the NpmConfig component does not add anything. In many configurations, this is causing an empty component, which is not doing anything. However it blocks the creation of a custom NpmConfig, causing a bad UX.

This change makes the `npmrc` property a lazy getter, so the component is only added when actually required.

Fixes #2638
Fixes #2635

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.