fix(cli): pass through --help flag for tasks that receive args (#3046)
A common use case for tasks is to act as proxies to other CLIs/tools (e.g. pip, poetry, etc). The current behavior of displaying projen's own help document for the `--help` flag for tasks prevents the proper passthrough of `--help` flags that are intended for the downstream tool (when `receiveArgs` is true).

The help document shown for tasks is fairly generic and can't be customized for tasks anyway.

I considered adding an option to the task itself (e.g. `disableHelp: boolean`) to control this behavior, but this current implementation makes more sense to me since `receiveArgs` should really mean all args. I realize that `--inspect` will still be captured by projen, but that one actually does make sense since a) it's not a standard flag on most tools and b) the return value is useful for understanding the task.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.