fix(object-file): allow boxed primitives, Set, Map and RegExp in values (#3016)
With current implementation ObjectFile (and all it subclasses) are throwing on encountering any perfectly valid Javascript boxed primitives in the object values (for example `new Sting('something')` or `new Boolean(something)`) even while those values may be stringified by `JSON.stringify` (the serialisation function that we are using) without any problem. This PR fixes that.

In addition, there are several JavaScript built-in objects that while having some limitations on JSII are very useful either for internal usage or for TS/JS-first usage of projen, however, those classes are missing built-in `.toJSON` method and so also resulting in throw from ObjectFile, particularly RegExp, Set and Map. This PR proposes to fix that by resolving RegExp (without flags) into its string source, Set into an array, and Map into an object. Note, that if we have a subclass of any of those built-in types that implements custom `.toJSON` then it will continue to work as with current version (see added tests).

As a simple demonstration of value of stringifiying `Set` I've included a small commit into this PR that switches implementation of ESLint fields that are expected to have non-repeated values (`plugins` and `extends`) from Array to Set using the fix implemented in this PR.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.