chore: fix CONTRIBUTING.md referring to wrong default branch (#3128)
### Description of changes

In the Release Versioning paragraph the default branch which this works from now is renamed to `main` and we want the document to reflect that. 



#### Fixes

* fix: changed master->main in version bump explainer paragraph

---
Thank you for your open source project. 
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.