feat: `compile` now automatically runs `pre-compile` and `post-compile`
This PR adds a dependency system to tasks; it also adds a reverse
dependency system (an "implies" system). This means that tasks
can now require other tasks to run before and after themselves.

The motivation for this was that when you have a `pre-compile` task, you
*always* want this task to run when you run `npx projen compile`; but in
the old system, the only way to run `pre-compile` was to run the `npx
projen build` task instead, which does a lot more work which is not
always desired.

When the same task is being requested in a single run of `projen`
(for example, if `pre-compile` is requested to be run both
as a dependency of `compile`, as well as implied by the `build`
task) duplicate invocations will be skipped.