fix(node): packaging bundled dependencies fails in pnpm projects (#3156)
When using `pnpm` as a package manager, bundled dependencies are hard linked and if using `npm pack`, these dependencies are not correctly included in the bundle. `pnpm pack` handles this by correctly including these hard linked packages into the output bundle.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.