feat(node): allow enforcement of licenses used by dependencies (#2614)
In some organizations only certain licenses are allowed to be used or a package want's to ensure it only depends on certain licenses.

This change adds a new feature to check the used licenses.
It will fail the build if any prohibited or not allowed license is encountered.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.