fix: WorkflowSteps enables completely consistent actions/checkout and setGitIdentity in Workflows (#3204)
Fixes https://github.com/projen/projen/issues/3203

Placed a bit of an opinion on WorkflowSteps with a common interface, please weigh in on how you like the pattern.  I moved an unrelated Step to the new interface in order to demonstrate.

This is non-breaking PR, but does include a deprecation, since I moved something.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.