feat(release): subproject releases (#3208)
Resolves https://github.com/projen/projen/issues/3172

This is non-breaking, but some opinions made:
- Releases for each project should be a separate workflow, named after the project in the case of a subproject
- For the base release workflow job (creation of artifact), the default working-directory should be set to the relative outdir of the subproject, with an exception made for the install command, which should be run from the topLevelParent of the subproject outdir.
- The folder that the artifact is published from, for a subproject, should be the subproject's outdir + artifactDirectory.
- Perhaps we should break TaskWorkflow, but I just decomposed it.
- Settings - Nothing has changed default-wise, and as such, in order to release subprojects, you must manually set ~~`github` to true~~ (@mrgrain comments) and `release` to true.

Snapshots have changed, but it has been verified that the default release behavior for NON-subprojects has remained identical.

I attempted to test this on a monorepo, but the package linking process combined with projen's automated nature proved too difficult, but based on the unit tests, I believe it would satisfy.

Open to more testing ideas, but wanted to ensure the opinions formed are the right ones.

I considered writing a recursive `releaseTagPrefix` conflict checker, but it seemed outside the scope, and should be another request if desired.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.