chore: new docusaurus-based documentation site (#3244)
This branch introduces the new /docusaurus directory which contains a new set of docs. These docs were written mostly from scratch to cover what we felt was the best starting point in using projen. However, these are recognized by the contributors as not being complete. We could probably spend another couple of months trying to fill them out better but felt we have hit a good point in the basic framework and content that we can merge this to upstream and let the community continue contributions.

A preview of this is available from our github pages link:

https://defiance-digital.github.io/projen/

Additionally, the images used are either manually created by us and we're not claiming any copyrights on them, or they are AI generated.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.