chore: Improve readability of project types on docs Introduction page (#3274)
# Objective
In this small PR, I am proposing that we update the doc's [introduction](https://projen.io/docs/introduction/) by breaking up the project types with headers. This will make it easier for developers to scan the list and identify the relevant project types.

## Before:
![image](https://github.com/projen/projen/assets/1624443/6b02f220-d0a2-42b6-9bc2-d231c99f1a32)

## After:
![image](https://github.com/projen/projen/assets/1624443/e2d99869-6072-44a5-8989-7173583a6162)


---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.