fix(jest): build fails when running with an old (<29) version of jest that is not explicitly specified as `jestVersion` (#3276)
In #3201, we started generating a new config to deal with an upcoming deprecation of some configuration setting in `ts-jest`.

However, this new configuration now breaks when running with an *old* version of `ts-jest`, showing the following:

```
    Unable to process '***', please make sure that `outDir` in your tsconfig is neither `''` or `'.'`. You can also configure Jest config option `transformIgnorePatterns` to inform `ts-jest` to transform ***
```

The component would deal with versions correctly if a specific version of `jest` was requested via its component properties, but if no `jest` version was requested it would assume it had the most recent version (note that we're using the `jest` version a proxy for the `ts-jest` version).

Add a feature where we check the actual version of `ts-jest` that is installed (by looking at `package.json`), rather than hoping the user tells us what version they're installing. This will make the new feature robust in the face of old dependencies.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.