fix(release): release workflow attempts to publish packages for already existing versions (#3252)
The workflow now checks if the to-be-published version already exists as a tag on the repository.
If the tag is found, we don't attempt to publish.

This previously worked, because all supported registries treat versions as immutable and refuse to publish over an existing version. With this change, no unnecessary publish attempts are made.

Resolves #2921 
Possible fix for https://github.com/projen/projen/issues/2763

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.