fix(release): workflow is failing because of invalid bash command (#3285)
A stray `'` made it's way into the codebase in #3252, causing the release workflow to fail.

This is currently blocking the release of projen. Which is good, because it means the broken code did not get published.
I don't know of a better way to catch these kind of mistakes. =/

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.