fix(project): `runTaskCommand()` ignores `projenCommand` (#3293)
The existing implementation of `runTaskCommand()` on `Project` has a bug that is causing it ignore the value of `projenCommand`.

While looking at fixing this, I noticed a bunch of inconsistency in this area. For example `NodeProject` has its own implementation of `runTaskCommand()` and `NodePackage` has a custom `projenCommand`, which I deprecated. There's also some inconsistency of when `projenCommand` is used directly vs. when `runTaskCommand()` is used and both have a different default value.

I don't have time to sort out this mess right now, because both "easy fixes" have some wider implications:
- Making `npx projen@<version>` the default for `projenCommand` feels safer, but would cause all generated files to update the header all the time
- Using `npx projen` in `runTaskCommand()` seems less safe for languages that don't have the projen version locked in a lockfile.

Thus I have focused on fixing the bug at hand. This is a low hanging fruit albeit the implementation is a bit akward.


---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.