fix(jsii): compile fails with `error TS1110: Type expected` due to incompatible `@types/node` version (#3331)
Fixes #3324

`@types/node@16.18.78` is the last `^16` version that works.
`@types/node^18.11.19` is the last `^18` version that works.

Also tested `@types/node^20` but these never worked, so I'm falling back to ^18 for them and log a warning.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.