feat(node): update aws-actions/configure-aws-credentials@v2 to v4 (#3329)
Fixes #2932 

As mentioned in the issue, I have updated the action from v2 to v4.

Now, the tests wouldn't pass locally before I made the change, and it appears to be related to #3324 
It seems the same error is occurring in recent builds as well.
![image](https://github.com/projen/projen/assets/12206103/5dcdd9e3-1865-4a5a-aebd-5a3201c3120a)

After implementing my change, I reran the tests and confirmed that the same test failed. Based on that, I assume my change hasn't affected anything else.

Also, a test projen project passes and behaves correctly.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.