feat(java): support for parent POM in maven projects (#3326)
Maven supports Parent Projects[0] which is very useful for controlling configuration of things like plugins across a set of projects. One of the biggest examples of this is with the spring boot maven parent[1] which helps setup plugins and provides versions to the spring boot dependencies underneath to keep them in line. 

[0] - https://maven.apache.org/guides/introduction/introduction-to-the-pom.html#project-inheritance
[1] - https://docs.spring.io/spring-boot/docs/current/reference/html/getting-started.html#getting-started.first-application.dependencies.maven

Fixes #

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.