fix(gitignore): paths separators are inconsistent for cross platform development (#3325)
Related to #3284

The issue could be related to more components, and this PR doesn't fully resolve it.

Normalize gitignore paths to be consistent across platforms. This avoids the diff shown between Windows (\/) and Linux (/) directory separators.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.