fix(github): invalid commit for tibdex/github-app-token action (#3341)
In PR #3338 we attempted to update the version of the `tibdex/github-app-token` GitHub Action to a later version.
We incorrectly picked the latest commit from the main branch of this action. However that's not how most GitHub Actions work. Instead they a tag a dangling commit that has been specially created with the required build artifacts.

This PR we are updates the hash to the correct working version as referenced from the release page: https://github.com/tibdex/github-app-token/tree/v2.1.0

Fixes: https://github.com/projen/projen/issues/3340

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.